requirements.txt
Flask
opencv-python
pytesseract
pillow
numpy
imutils